backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-genai==0.3.0  # ¡NUEVA!
python-multipart==0.0.6
python-dotenv==1.0.0
pypdf2==3.0.1
pillow==10.1.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1